+++ docs/paper/paper.bib
@@ -15,6 +15,7 @@
    journal={Journal of Choice Modelling},
    publisher={Elsevier BV},
    author={Brathwaite, Timothy and Walker, Joan L.},
+   doi = {https://doi.org/10.1016/j.jocm.2018.01.002},
    year={2018},
    month=dec,
    pages={78–112}
@@ -25,12 +26,14 @@
       author={Tianyu Du and Ayush Kanodia and Susan Athey},
       year={2023},
   journal={arXiv preprint arXiv:{2304.01906}},
+  doi = {https://doi.org/10.48550/arXiv.2304.01906},
 }
 
 @article{Aouad:2023,
   title={Representing random utility choice models with neural networks},
   author={Aouad, Ali and D{\'e}sir, Antoine},
   journal={arXiv preprint arXiv:2207.12877},
+  doi = {https://doi.org/10.48550/arXiv.2207.12877},
   year={2022}
 }
 
@@ -43,6 +46,7 @@ year = {2022},
 issn = {0191-2615},
 author = {Yafei Han and Francisco Camara Pereira and Moshe Ben-Akiva and Christopher Zegras},
 keywords = {Discrete choice models, Neural networks, Taste heterogeneity, Interpretability, Utility specification, Machine learning, Deep learning},
+doi = {https://doi.org/10.1016/j.trb.2022.07.001},
 abstract = {Discrete choice models (DCMs) require a priori knowledge of the utility functions, especially how tastes vary across individuals. Utility misspecification may lead to biased estimates, inaccurate interpretations and limited predictability. In this paper, we utilize a neural network to learn taste representation. Our formulation consists of two modules: a neural network (TasteNet) that learns taste parameters (e.g., time coefficient) as flexible functions of individual characteristics; and a multinomial logit (MNL) model with utility functions defined with expert knowledge. Taste parameters learned by the neural network are fed into the choice model and link the two modules. Our approach extends the L-MNL model (Sifringer et al., 2020) by allowing the neural network to learn the interactions between individual characteristics and alternative attributes. Moreover, we formalize and strengthen the interpretability condition — requiring realistic estimates of behavior indicators (e.g., value-of-time, elasticity) at the disaggregated level, which is crucial for a model to be suitable for scenario analysis and policy decisions. Through a unique network architecture and parameter transformation, we incorporate prior knowledge and guide the neural network to output realistic behavior indicators at the disaggregated level. We show that TasteNet-MNL reaches the ground-truth model’s predictability and recovers the nonlinear taste functions on synthetic data. Its estimated value-of-time and choice elasticities at the individual level are close to the ground truth. In contrast, exemplary logit models with misspecified systematic utility lead to biased parameter estimates and lower prediction accuracy. On a publicly available Swissmetro dataset, TasteNet-MNL outperforms benchmarking MNLs and Mixed Logit model’s predictability. It learns a broader spectrum of taste variations within the population and suggests a higher average value-of-time. Our source code is available for research and application.}
 }
 
@@ -50,6 +54,7 @@ abstract = {Discrete choice models (DCMs) require a priori knowledge of the util
   title={RUMBoost: Gradient Boosted Random Utility Models},
   author={Salvad{\'e}, Nicolas and Hillel, Tim},
   journal={arXiv preprint arXiv:2401.11954},
+  doi={https://doi.org/10.48550/arXiv.2401.11954},
   year={2024}
 }
 
[…]
