### 2024-06-14 #6882 @tmigot
> 👋 @tmigot - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Sure. I can review it.
