### 2024-06-17 #6882 @samuelduchesne
> 👋 @samuelduchesne - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I can review it as well.
