### 2024-06-17 e21e251
Update README.md (#107)
* Update README.md

* FIX: lint

---------

Co-authored-by: VincentAuriau <auriau.vincent@gmail.com>
