### 2024-06-17 #6899 @editorialbot
[REVIEW]: Choice-Learn: Large-scale choice modeling for operational contexts through the lens of machine learning
**Submitting author:** @VincentAuriau (<a href="http://orcid.org/0009-0002-9640-2639">Vincent Auriau</a>)
**Repository:** https://github.com/artefactory/choice-learn
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
