### 2024-06-17 #6899 @VincentAuriau
Hello,
Thank you for reviewing  !
A few experiments are detailed in the paper and can be reproduced with the notebooks [here](https://github.com/artefactory/choice-learn/tree/joss/docs/paper).
