### 2024-06-17 a9044ac
attempt at reducing word count
