### 2024-06-18 131e221
FIX: ArrayStorage when directly instantiated from ndarray
