### 2024-06-19 54a9b50
ENH: better handling when features / availabilities are indexed with FeaturesStorage
