### 2024-06-21 6bcb634
ADD: ArrayStorage indexer & catching KeyErrors for custom messaging
