### 2024-06-21 180f77f
ENH: ArrayStorage follows common structure with Indexers
