### 2024-06-21 b87064e
ENH: harmonize Error raised with missing key
