### 2024-06-21 d786675
ADD: KeyError test for all Storages
