### 2024-06-24 3618b37
Paper length reduction (#109)
* ADD: small enhancements

* FIX: typo

* ENH/FIX: plots
