### 2024-06-24 b62f9eb
ENH/FIX: Notebooks & Documentation (#110)
* ENH: few modifications in doc

* ENH: Documentation Landing page
