+++ docs/index.md
@@ -1,12 +1,23 @@
 # Welcome to the choice-learn documentation!
 
-A toolbox for choice-modeling
+<div align="center">
+
+<img src="https://github.com/artefactory/choice-learn/blob/main/docs/illustrations/logos/logo_choice_learn.png" width="300">
+
+*Large-scale choice modeling through the lens of machine learning*
 
 
-Choice-Learn is a Python package designed to help you build discrete choice models.
-The package provides ready to use datasets and different models from the litterature. It also provides a lower level use if you want to customize any model or create your own from scratch. In particular you will find smart datasets handling to limit RAM usage and different structure commons to any choice model.
+| [<img src="https://upload.wikimedia.org/wikipedia/commons/thumb/c/c2/GitHub_Invertocat_Logo.svg/langfr-400px-GitHub_Invertocat_Logo.svg.png" alt="drawing" width="40"/>](https://github.com/artefactory/choice-learn) | If you are not coming from GitHub, check it out first [here](https://github.com/artefactory/choice-learn)! |
+|---|---|
 
-Choice-Learn uses NumPy and pandas as data backend engines and TensorFlow for models.
+</div>
+<br>
+<br>
+
+Choice-Learn is a Python package designed to help you estimate discrete choice models and use them (e.g., assortment optimization plug-in).
+The package provides ready-to-use datasets and models from the litterature. It also provides a lower level use if you wish to customize any choice model or create your own from scratch. Choice-Learn efficiently handles data with the objective to limit RAM usage. It is made particularly easy to estimate choice models with your own, large datasets.
+
+Choice-Learn uses NumPy and pandas as data backend engines and TensorFlow for models.els.
 
 In this documentation you will find examples to be quickly getting started as well as some more in-depth example.
 
@@ -17,13 +28,13 @@ Here is a quick overview of the different functionalities offered by Choice-Lear
 ### Data
 - [Custom data handling](./reference/data/references_choice_dataset.md) for choice datasets with possible memory usage optimizations
 - Some Open-Source ready-to use datasets are included within the datasets:
-  - [SwissMetro](./references/dataset/references_base.md)
-  - [ModeCanada](./references/dataset/references_base.md)
-  - The [Train](./references/dataset/references_base.md) dataset
-  - The [Heating](./references/dataset/references_base.md) & [Electricity](./references/dataset/references_base.md)datasets from Kenneth Train
-  - [Stated car preferences](./references/dataset/references_base.md)
-  - The [TaFeng](./references/dataset/references_tafeng.md) dataset from Kaggle
-  - The ICDM-2013 [Expedia](./references/dataset/references_expedia.md) dataset from Kaggle
+    - [SwissMetro](./references/dataset/references_base.md)
+    - [ModeCanada](./references/dataset/references_base.md)
+    - The [Train](./references/dataset/references_base.md) dataset
+    - The [Heating](./references/dataset/references_base.md) & [Electricity](./references/dataset/references_base.md)datasets from Kenneth Train
+    - [Stated car preferences](./references/dataset/references_base.md)
+    - The [TaFeng](./references/dataset/references_tafeng.md) dataset from Kaggle
+    - The ICDM-2013 [Expedia](./references/dataset/references_expedia.md) dataset from Kaggle
 
 ### Models
 - [Custom modelling](./notebooks/introduction/4_model_customization.md)
@@ -32,6 +43,7 @@ Here is a quick overview of the different functionalities offered by Choice-Lear
         - [Multinomial Logit](./references/models/references_simple_mnl.md)
         - [Conditional Logit](./references/models/references_clogit.md)
[…]
