### 2024-06-25 1bf3d3b
ENH: diverse improvements in the documentation
