+++ docs/tutorials.md
@@ -1,3 +1,4 @@
-Here are two tutorials to get started with the choice-learn package.
-In particular it shows how to create a ChoiceDataset. It's important as its the main object that will handle your data for the model.
-The other example shows a few example on how to handle ConditionalLogit with the package.
+Here are a few tutorials to get started with the choice-learn package.
+In particular it shows how handle data and get started with modelling.
+
+You will find  mode in-depth examples in the How-To section.
