### 2024-06-26 8f1cb66
ENH: FeaturesStorage IDs checked automatically
