### 2024-06-26 6ebb887
ENH: Way faster FeaturesStorage instantiation in some edge cases
