### 2024-06-26 8e1398b
ENH: use numpy function instead of rewriting
