+++ choice_learn/data/indexer.py
@@ -2,6 +2,7 @@
 
 import logging
 from abc import abstractmethod
+
 import numpy as np
 
 
