### 2024-06-26 21b71d5
ENH: trainable_weights as abstractmethod & property
