### 2024-06-26 131f6b5
ENH: small enhancements in README
