### 2024-06-26 b1e0d2d
ADD: simpleMNL fit as test
