+++ requirements.txt
@@ -2,5 +2,5 @@
 numpy==1.24.3
 pandas==1.5.3
 tensorflow==2.13.0
-tensorflow_probability==0.20.1
+tensorflow_probability==0.22.1
 tqdm==4.65.0
