+++ requirements.txt
@@ -1,6 +1,6 @@
 -e .
 numpy==1.24.3
 pandas==1.5.3
-tensorflow==2.13.0
+tensorflow==2.14.0
 tensorflow_probability==0.22.1
 tqdm==4.65.0
