+++ pyproject.toml
@@ -79,7 +79,7 @@ classifiers = [
 python = "^3.9.0"
 numpy = "^1.24.3"
 pandas = "^1.5.3"
-tensorflow = "^2.11.0"
+tensorflow = "^2.14.0"
 tensorflow-probability = "^0.22.1"
 tqdm = "^4.0.0"
 ortools = { version = "^9.6", optional = true }
