### 2024-07-01 d16b326
ADD: few SimpleMNL tests
