+++ tests/integration_tests/models/test_conditional_logit.py
@@ -2,8 +2,8 @@
 
 import tensorflow as tf
 
-from choice_learn.datasets import load_modecanada, load_swissmetro
-from choice_learn.models import ConditionalLogit, SimpleMNL
+from choice_learn.datasets import load_modecanada
+from choice_learn.models import ConditionalLogit
 
 
 def test_mode_canada_gt():
@@ -33,16 +33,6 @@ def test_mode_canada_gt():
     gt_model.instantiate(canada_dataset)
 
     gt_model.trainable_weights = gt_weights
-    assert (gt_model.evaluate(canada_dataset) * len(canada_dataset)) <= 1874.4, gt_model.evaluate(
-        canada_dataset
-    ) * len(canada_dataset)
-    assert (gt_model.evaluate(canada_dataset) * len(canada_dataset)) >= 1874.1
-
-
-def test_simple_mnl_fit():
-    """Tests that SimpleMNL can fit."""
-    dataset = load_swissmetro()
-    model = SimpleMNL()
-    model.fit(dataset)
-    model.evaluate(dataset)
-    assert model.evaluate(dataset) < 1.0
+    total_nll = gt_model.evaluate(canada_dataset) * len(canada_dataset)
+    assert total_nll <= 1874.4, f"Got NLL: {total_nll}"
+    assert total_nll >= 1874.1, f"Got NLL: {total_nll}"
