### 2024-07-01 1af1e64
Hotfix/storage (#111)
ADD: Custom KeyErrors with FeaturesStorage
ADD: ArrayStorageIndexer to follow common structure

ENH: ChoiceDataset does not automatically checks all IDs of FeaturesStorage, it is moved to a specific function that can be manually called

FIX: ArrayStorage when directly instantiated from ndarray
