+++ choice_learn/data/choice_dataset.py
@@ -1,4 +1,5 @@
 """Main classes to handle assortment data."""
+
 import logging
 
 import numpy as np
@@ -99,7 +100,8 @@ class ChoiceDataset(object):
                         if len(sub_features[0]) != len(sub_names):
                             raise ValueError(
                                 f"""{sub_k}-th given shared_features_by_choice and
-                                shared_features_by_choice_names shapes do not match"""
+                                shared_features_by_choice_names shapes do not match:
+                                {len(sub_features[0])} and {len(sub_names)}."""
                             )
                 # In this case names are missing, still transform it as a tuple
                 else:
@@ -141,7 +143,7 @@ class ChoiceDataset(object):
                 if np.array(sub_features).ndim == 1:
                     # check features_by_ids
                     logging.info(
-                        "feature of dimension 1 detected -  a FeatureByIDs MUst be provided"
+                        "feature of dimension 1 detected -  a FeatureByIDs MUST be provided"
                     )
                     for fbid in features_by_ids:
                         if fbid.name == sub_names[0]:
@@ -155,8 +157,10 @@ class ChoiceDataset(object):
 
                 elif len(sub_features[0][0]) != len(sub_names):
                     raise ValueError(
-                        f"""{sub_k}-th given items_features_by_choice and
-                        items_features_by_choice_names shapes do not match"""
+                        f"""{sub_k}-th given items_features_by_choice with names
+                        {sub_names} and
+                        items_features_by_choice_names shapes do not match:
+                        {len(sub_features[0][0])} and {len(sub_names)}."""
                     )
             self._return_items_features_by_choice_tuple = True
 
@@ -357,6 +361,7 @@ class ChoiceDataset(object):
             self.shared_features_by_choice_map,
             self.items_features_by_choice_map,
         ) = self._build_features_by_ids()
+        self.check_features_by_ids()
 
         # self.n_choices = len(self.choices)
 
@@ -364,7 +369,6 @@ class ChoiceDataset(object):
         self._check_dataset()  # Should handle alone if np.arrays are squeezed
         self._return_types = self._check_types()
         self._check_names()
-
         # Build .iloc method
         self.indexer = ChoiceDatasetIndexer(self)
 
@@ -436,10 +440,13 @@ class ChoiceDataset(object):
                                     {feature_by_id.name}"""
                                 )
 
-                                unique_values = np.unique(self.shared_features_by_choice[i][:, j])
+                                # We test a subset of IDs
+                                test_values = self.shared_features_by_choice[i][:, j][
+                                    np.arange(0, len(self.choices), 10)
+                                ]
                                 try:
-                                    for val in unique_values:
-                                        feature_by_id.batch[unique_values]
+                                    for val in test_values:
+                                        feature_by_id.batch[val]
                                 except KeyError:
                                     raise ValueError(
                                         f"""Key {val} in Shared Feature {column_name}
@@ -459,29 +466,65 @@ class ChoiceDataset(object):
                                     f"""Feature by ID found for items_features_by_choice:
                                     {feature_by_id.name}"""
                                 )
+
+                                # We test a subset of the IDs
                                 if self.items_features_by_choice[i].ndim == 1:
-                                    unique_values = np.unique(self.items_features_by_choice[i])
+                                    test_values = self.items_features_by_choice[i][
+                                        np.arange(0, len(self.choices), 10)
+                                    ]
                                 else:
-                                    unique_values = np.unique(
-                                        self.items_features_by_choice[i][:, :, k]
-                                    )
+                                    test_values = self.items_features_by_choice[i][:, :, k][
+                                        np.arange(0, len(self.choices), 10)
[…]
