+++ tests/unit_tests/data/test_storage.py
@@ -1,11 +1,11 @@
-"""Test the store module."""
+"""Test the storage module."""
 import numpy as np
 import pandas as pd
 
-from choice_learn.data.storage import FeaturesStorage
+from choice_learn.data.storage import ArrayStorage, DictStorage, FeaturesStorage, OneHotStorage
 
 
-def test_len_store():
+def test_len_storage():
     """Test the __len__ method of Storage."""
     features = {"customerA": [1, 2], "customerB": [4, 5], "customerC": [7, 8]}
     storage = FeaturesStorage(
@@ -15,8 +15,8 @@ def test_len_store():
     assert storage.shape == (3, 2)
 
 
-def test_get_store_element():
-    """Test the _get_store_element method of Store."""
+def test_get_storage_element():
+    """Test the _get_storage_element method of storage."""
     features = {"customerA": [1, 2], "customerB": [4, 5], "customerC": [7, 8]}
     storage = FeaturesStorage(
         values=features, values_names=["age", "income", "children_nb"], name="customers"
@@ -25,8 +25,8 @@ def test_get_store_element():
     assert (storage.get_element_from_index([0, 1, 2]) == np.array([[1, 2], [4, 5], [7, 8]])).all()
 
 
-def test_store_batch():
-    """Test the batch method of Store."""
+def test_storage_batch():
+    """Test the batch method of storage."""
     features = {"customerA": [1, 2], "customerB": [4, 5], "customerC": [7, 8]}
     storage = FeaturesStorage(
         values=features, values_names=["age", "income", "children_nb"], name="customers"
@@ -38,8 +38,8 @@ def test_store_batch():
     ).all()
 
 
-def test_featuresstore_instantiation():
-    """Test the instantiation of FeaturesStore."""
+def test_featuresstorage_instantiation():
+    """Test the instantiation of Featuresstorage."""
     features = {"customerA": [1, 2], "customerB": [4, 5], "customerC": [7, 8]}
     storage = FeaturesStorage(
         values=features, values_names=["age", "income", "children_nb"], name="customers"
@@ -56,8 +56,8 @@ def test_featuresstore_instantiation():
         ).all()
 
 
-def test_featuresstore_instantiation_indexless():
-    """Test the instantiation of FeaturesStore."""
+def test_featuresstorage_instantiation_indexless():
+    """Test the instantiation of Featuresstorage."""
     features = [[1, 2, 3], [4, 5, 6], [7, 8, 9]]
     ids = ["customerA", "customerB", "customerC"]
 
@@ -76,8 +76,8 @@ def test_featuresstore_instantiation_indexless():
         ).all()
 
 
-def test_featuresstore_instantiation_from_list():
-    """Test the instantiation of FeaturesStore."""
+def test_featuresstorage_instantiation_from_list():
+    """Test the instantiation of Featuresstorage."""
     features = [[1, 2, 3], [4, 5, 6], [7, 8, 9]]
 
     storage = FeaturesStorage(
@@ -90,8 +90,8 @@ def test_featuresstore_instantiation_from_list():
     ).all()
 
 
-def test_array_store_with_ids():
-    """Test the instantiation of FeaturesStore."""
+def test_array_storage_with_ids():
+    """Test the instantiation of Featuresstorage."""
     features = [[1, 2, 3], [4, 5, 6], [7, 8, 9]]
 
     storage = FeaturesStorage(
@@ -107,8 +107,8 @@ def test_array_store_with_ids():
     ).all()
 
 
-def test_array_store_with_mixed_ids():
[…]
