### 2024-07-01 24102aa
Enh/mkdocs (#114)
* ENH: diverse improvements in the documentation

* ENH: small enhancements in README
