+++ docs/index.md
@@ -1,4 +1,4 @@
-# Welcome to the choice-learn documentation!
+# Welcome to the Choice-Learn documentation!
 
 <center>
 <img src="illustrations/logos/logo_choice_learn.png" width="300">
@@ -12,12 +12,12 @@
 
 <center>
 
-| [<img src="https://upload.wikimedia.org/wikipedia/commons/thumb/c/c2/GitHub_Invertocat_Logo.svg/langfr-400px-GitHub_Invertocat_Logo.svg.png" alt="drawing" width="40"/>](https://github.com/artefactory/choice-learn) | If you are not coming from GitHub, check it out first [here](https://github.com/artefactory/choice-learn)! |
+| [<img src="https://upload.wikimedia.org/wikipedia/commons/thumb/c/c2/GitHub_Invertocat_Logo.svg/langfr-400px-GitHub_Invertocat_Logo.svg.png" alt="drawing" width="30"/>](https://github.com/artefactory/choice-learn) | If you are not coming from GitHub, [check it out](https://github.com/artefactory/choice-learn) |
 |-|-|
+| [<img src="https://upload.wikimedia.org/wikipedia/commons/8/87/PDF_file_icon.svg" alt="drawing" width="25"/>](https://github.com/artefactory/choice-learn/actions/runs/9642058728/artifacts/1630710678) | You can also read our [academic paper](https://github.com/artefactory/choice-learn/actions/runs/9642058728/artifacts/1630710678) ! |
 
 </center>
 
-<br>
 <br>
 
 Choice-Learn is a Python package designed to help you estimate discrete choice models and use them (e.g., assortment optimization plug-in).
@@ -32,15 +32,15 @@ In this documentation you will find examples to be quickly getting started as we
 Here is a quick overview of the different functionalities offered by Choice-Learn. Further details are given in the rest of the documentation.
 
 ### Data
-- [Custom data handling](./reference/data/references_choice_dataset.md) for choice datasets with possible memory usage optimizations
+- [Custom data handling](./references/data/references_choice_dataset.md) for choice datasets with possible memory usage optimizations
 - Some Open-Source ready-to use datasets are included within the datasets:
-    - [SwissMetro](./references/dataset/references_base.md)
-    - [ModeCanada](./references/dataset/references_base.md)
-    - The [Train](./references/dataset/references_base.md) dataset
-    - The [Heating](./references/dataset/references_base.md) & [Electricity](./references/dataset/references_base.md)datasets from Kenneth Train
-    - [Stated car preferences](./references/dataset/references_base.md)
-    - The [TaFeng](./references/dataset/references_tafeng.md) dataset from Kaggle
-    - The ICDM-2013 [Expedia](./references/dataset/references_expedia.md) dataset from Kaggle
+    - [SwissMetro](./references/datasets/references_base.md)
+    - [ModeCanada](./references/datasets/references_base.md)
+    - The [Train](./references/datasets/references_base.md) dataset
+    - The [Heating](./references/datasets/references_base.md) & [Electricity](./references/dataset/references_base.md)datasets from Kenneth Train
+    - [Stated car preferences](./references/datasets/references_base.md)
+    - The [TaFeng](./references/datasets/references_tafeng.md) dataset from Kaggle
+    - The ICDM-2013 [Expedia](./references/datasets/references_expedia.md) dataset from Kaggle
 
 ### Models
 - [Custom modelling](./notebooks/introduction/4_model_customization.md)
@@ -130,3 +130,7 @@ The use of this software is under the MIT license, with no limitation of usage, 
 ### Affiliations
 
[…]
