### 2024-07-01 610165a
Fix/tf tfp compat (#115)
* ENH: small enhancements in README & Doc

* ENH: higher TF & TFP version

* ADD: few SimpleMNL tests
