+++ tests/integration_tests/models/test_conditional_logit.py
@@ -33,7 +33,6 @@ def test_mode_canada_gt():
     gt_model.instantiate(canada_dataset)
 
     gt_model.trainable_weights = gt_weights
-    assert (gt_model.evaluate(canada_dataset) * len(canada_dataset)) <= 1874.4, gt_model.evaluate(
-        canada_dataset
-    ) * len(canada_dataset)
-    assert (gt_model.evaluate(canada_dataset) * len(canada_dataset)) >= 1874.1
+    total_nll = gt_model.evaluate(canada_dataset) * len(canada_dataset)
+    assert total_nll <= 1874.4, f"Got NLL: {total_nll}"
+    assert total_nll >= 1874.1, f"Got NLL: {total_nll}"
