### 2024-07-02 dcfb380
FIX: subset ChoiceDataset with availabilities as FbID
