### 2024-07-03 e86530c
ENH: better html for logos
