### 2024-07-03 1a26eee
ADD: right links with logos
