+++ choice_learn/data/choice_dataset.py
@@ -1233,6 +1233,10 @@ class ChoiceDataset(object):
                                 ].batch[shared_features_by_choice[tuple_index][:, feature_index]]
                             )
                             feat_ind_min = feature_index + 1
+                        if feat_ind_min != shared_features_by_choice[tuple_index].shape[1]:
+                            unstacked_feat.append(
+                                shared_features_by_choice[tuple_index][:, feat_ind_min:]
+                            )
                         mapped_features.append(np.concatenate(unstacked_feat, axis=1))
                     else:
                         mapped_features.append(shared_features_by_choice[tuple_index])
@@ -1268,6 +1272,10 @@ class ChoiceDataset(object):
                                     ]
                                 )
                                 feat_ind_min = feature_index + 1
+                            if feat_ind_min != items_features_by_choice[tuple_index].shape[2]:
+                                unstacked_feat.append(
+                                    shared_features_by_choice[tuple_index][:, :, feat_ind_min:]
+                                )
                             mapped_features.append(np.concatenate(unstacked_feat, axis=2))
                     else:
                         mapped_features.append(items_features_by_choice[tuple_index])
