+++ choice_learn/data/indexer.py
@@ -405,11 +405,17 @@ class ChoiceDatasetIndexer(Indexer):
                                     ].keys()
                                 )
                             ):
-                                unstacked_feat.append(
-                                    items_features_by_choice[tuple_index][
-                                        :, :, feat_ind_min:feature_index
-                                    ]
-                                )
+                                if feat_ind_min != feature_index:
+                                    unstacked_feat.append(
+                                        shared_features_by_choice[tuple_index][
+                                            :, feat_ind_min:feature_index
+                                        ]
+                                    )
+                                    unstacked_feat.append(
+                                        items_features_by_choice[tuple_index][
+                                            :, :, feat_ind_min:feature_index
+                                        ]
+                                    )
                                 unstacked_feat.append(
                                     self.choice_dataset.items_features_by_choice_map[tuple_index][
                                         feature_index
