### 2024-07-03 2c1ba45
ADD: trainable weights as @propery
