### 2024-07-03 069cd7d
ADD: tests with latest changes
