### 2024-07-04 9d9288a
Small enhancements to the ReadMe (#120)
