### 2024-07-11 6a2139a
FIX: rumnet regularization & evaluate
