### 2024-07-17 068239f
Enh/tqdm bar (#128)
* ENH: description of tqdm bars with losses
