### 2024-07-17 d7f9056
ADD: coverage in PR CI from example
