+++ pyproject.toml
@@ -87,6 +87,7 @@ gurobipy = { version = "^11.0", optional = true }
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.3.2"
+pytest-cov = "^4.1.0"
 pre-commit = "^3.3"
 ipykernel = "^6.9"
 nbstripout = "^0.7"
