### 2024-07-17 60058c3
ADD: pytest-cov in requirements-dev.txt
