+++ pyproject.toml
@@ -87,12 +87,11 @@ gurobipy = { version = "^11.0", optional = true }
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.3.2"
-pytest-cov = "^4.0.0"
 pre-commit = "^3.3"
 ipykernel = "^6.9"
 nbstripout = "^0.7"
 ruff = "^0.1.2"
-pytest-cov = "^4.1"
+pytest-cov = "^4.0"
 codespell = "^2.2"
 python-markdown-math = "^0.8"
 
