+++ pyproject.toml
@@ -91,7 +91,7 @@ pre-commit = "^3.3"
 ipykernel = "^6.9"
 nbstripout = "^0.7"
 ruff = "^0.1.2"
-pytest-cov = "^4.0"
+pytest-cov = "^4.1"
 codespell = "^2.2"
 python-markdown-math = "^0.8"
 
