### 2024-07-17 dc103bc
ADD: coverage in PR CI from example (#130)
GH Actions automatically comments a PyTest coverage report in Pull Requests
