### 2024-07-17 6d8efd8
ENH: raw paths to imgs
