### 2024-07-19 bb87e2b
ENH: dataset is now always choice_dataset
