### 2024-07-19 d49de22
ADD: Basic tests for baseline models
