+++ choice_learn/models/nested_logit.py
@@ -1,4 +1,5 @@
 """Implementation of the Nested Logit model."""
+
 import logging
 
 import numpy as np
@@ -346,16 +347,16 @@ class NestedLogit(ChoiceModel):
 
         self.coefficients = coefficients
 
-    def _store_dataset_features_names(self, dataset):
+    def _store_dataset_features_names(self, choice_dataset):
         """Register the name of the features in the dataset. For later use in utility computation.
 
         Parameters
         ----------
-        dataset : ChoiceDataset
+        choice_dataset : ChoiceDataset
             ChoiceDataset used to fit the model.
         """
-        self._shared_features_by_choice_names = dataset.shared_features_by_choice_names
-        self._items_features_by_choice_names = dataset.items_features_by_choice_names
+        self._shared_features_by_choice_names = choice_dataset.shared_features_by_choice_names
+        self._items_features_by_choice_names = choice_dataset.items_features_by_choice_names
 
     def compute_batch_utility(
         self,
@@ -750,7 +751,7 @@ class NestedLogit(ChoiceModel):
         self.instantiate(choice_dataset)
 
         fit = super()._fit_with_lbfgs(
-            dataset=choice_dataset,
+            choice_dataset=choice_dataset,
             sample_weight=sample_weight,
             **kwargs,
         )
@@ -758,12 +759,12 @@ class NestedLogit(ChoiceModel):
             self.report = self.compute_report(choice_dataset)
         return fit
 
-    def compute_report(self, dataset):
+    def compute_report(self, choice_dataset):
         """Compute a report of the estimated weights.
 
         Parameters
         ----------
-        dataset : ChoiceDataset
+        choice_dataset : ChoiceDataset
             ChoiceDataset used for the estimation of the weights that will be
             used to compute the Std Err of this estimation.
 
@@ -774,7 +775,7 @@ class NestedLogit(ChoiceModel):
         """
         import tensorflow_probability as tfp
 
-        weights_std = self.get_weights_std(dataset)
+        weights_std = self.get_weights_std(choice_dataset)
         dist = tfp.distributions.Normal(loc=0.0, scale=1.0)
 
         names = []
@@ -803,12 +804,12 @@ class NestedLogit(ChoiceModel):
             },
         )
 
-    def get_weights_std(self, dataset):
+    def get_weights_std(self, choice_dataset):
         """Approximates Std Err with Hessian matrix.
 
         Parameters
         ----------
-        dataset : ChoiceDataset
+        choice_dataset : ChoiceDataset
             ChoiceDataset used for the estimation of the weights that will be
             used to compute the Std Err of this estimation.
 
@@ -830,7 +831,7 @@ class NestedLogit(ChoiceModel):
                     mw.append(w[:, index : index + _w.shape[1]])
                     index += _w.shape[1]
                 model.trainable_weights = mw
-                batch = next(dataset.iter_batch(batch_size=-1))
+                batch = next(choice_dataset.iter_batch(batch_size=-1))
                 utilities = model.compute_batch_utility(*batch)
 
                 batch_gammas = []
@@ -839,11 +840,11 @@ class NestedLogit(ChoiceModel):
                 else:
                     for i in range(len(self.items_to_nest)):
                         if model.items_to_nest[i] == -1:
-                            batch_gammas.append([tf.constant(1.0)] * len(dataset))
+                            batch_gammas.append([tf.constant(1.0)] * len(choice_dataset))
                         else:
                             batch_gammas.append(
                                 [model.trainable_weights[-1][0, model.items_to_nest[i]]]
-                                * len(dataset)
+                                * len(choice_dataset)
                             )
[…]
