### 2024-07-19 9fcc66e
FIX: path to images in mkdoc
