### 2024-07-19 883cb65
FIX: ReadMe example (#125)
* FIX: ReadMe example

* ENH: change example comments

* ENH: raw paths to imgs

* FIX: path to images in mkdoc
