+++ docs/index.md
@@ -125,37 +125,37 @@ The use of this software is under the MIT license, with no limitation of usage, 
 
 ### Affiliations
 
-This package has been developped within the [Artefact Research Center](https://www.artefact.com/data-consulting-transformation/artefact-research-center/) in collaboration with CentraleSupélec, université Paris-Saclay.
-
+Choice-Learn has been developed through a collaboration between researchers at the Artefact Research Center and the laboratory MICS from CentraleSupélec, Université Paris Saclay.
 
 <p align="center">
   <a href="https://www.artefact.com/data-consulting-transformation/artefact-research-center/">
-    <img src="./docs/illustrations/logos/logo_arc.png" height="60" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/logo_arc.png" height="60" />
   </a>
   &emsp;
   &emsp;
   <a href="https://www.artefact.com/">
-    <img src="docs/illustrations/logos/logo_atf.png" height="65" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/logo_atf.png" height="65" />
+
   </a>
 </p>
 
 <p align="center">
   <a href="https://www.universite-paris-saclay.fr/">
-    <img src="./docs/illustrations/logos/logo_paris_saclay.png" height="60" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/logo_paris_saclay.png" height="60" />
   </a>
   &emsp;
   &emsp;
   <a href="https://mics.centralesupelec.fr/">
-    <img src="docs/illustrations/logos/logo_CS.png" height="60" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/logo_CS.png" height="60" />
   </a>
   &emsp;
   &emsp;
   <a href="https://www.london.edu/">
-    <img src="docs/illustrations/logos/logo_lbs.jpeg" height="60" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/logo_lbs.jpeg" height="60" />
   </a>
   &emsp;
   &emsp;
   <a href="https://www.insead.edu/">
-    <img src="docs/illustrations/logos/logo_insead.png" height="60" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/logo_insead.png" height="60" />
   </a>
 </p>
