+++ README.md
@@ -1,7 +1,7 @@
 <div align="center">
 
 
-<img src="https://github.com/artefactory/choice-learn/blob/main/docs/illustrations/logos/logo_choice_learn.png" width="300">
+<img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/logo_choice_learn.png" width="300">
 
 *Large-scale choice modeling through the lens of machine learning*
 
@@ -110,48 +110,49 @@ Finally for pricing or assortment optimization, you need either Gurobi or OR-Too
 
 <p align="center">
   <a href="https://numpy.org/">
-    <img src="docs/illustrations/logos/numpy_logo.png" width="60" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/numpy_logo.png" width="60" />
   </a>
   &nbsp;
   &nbsp;
   <a href="https://pandas.pydata.org/">
-    <img src="docs/illustrations/logos/pandas_logo.png" width="60" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/pandas_logo.png" width="60" />
   </a>
   &nbsp;
   &nbsp;
   <a href="https://www.tensorflow.org">
-    <img src="docs/illustrations/logos/tf_logo.png" width="60" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/tf_logo.png" width="60" />
   </a>
   &nbsp;
   &nbsp;
   <a href="https://www.gurobi.com/">
-    <img src="docs/illustrations/logos/gurobi_logo.png" width="60" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/gurobi_logo.png" width="60" />
   </a>
   &nbsp;
   &nbsp;
   <a href="https://developers.google.com/optimization?hl=fr">
-    <img src="docs/illustrations/logos/or_tools_logo.png" width="70" />
+    <img src="https://raw.githubusercontent.com/artefactory/choice-learn/main/docs/illustrations/logos/or_tools_logo.png" width="70" />
   </a>
 </p>
 
 > :bulb: **Tip:** You can use the poetry.lock or requirements-complete.txt files with poetry or pip to install a fully predetermined and working environment.
 
 ## :trident: Usage
-Here is a short example of model parametrization to estimate a Conditional Logit on the SwissMetro dataset.
+Here is a short example of model parametrization to estimate a Conditional Logit on the ModeCanada dataset.
 
 ```python
 from choice_learn.data import ChoiceDataset
 from choice_learn.models import ConditionalLogit, RUMnet
+from choice_learn.datasets import load_modecanada
 
+transport_df = load_modecanada(as_frame=True)
 # Instantiation of a ChoiceDataset from a pandas.DataFrame
-# Only need to specify how the file is encoded:
 dataset = ChoiceDataset.from_single_long_df(df=transport_df,
                                             items_id_column="alt",
                                             choices_id_column="case",
                                             choices_column="choice",
                                             shared_features_columns=["income"],
                                             items_features_columns=["cost", "freq", "ovt", "ivt"],
-                                            choice_format="item_id")
+                                            choice_format="one_zero")
 
 # Initialization of the model
 model = ConditionalLogit()
@@ -167,7 +168,7 @@ model.add_coefficients(feature_name="income",
 model.add_coefficients(feature_name="ivt",
                        items_indexes=[0, 1, 2, 3])
 
[…]
