### 2024-07-22 a15fa4d
ADD: additional tests for baseline models
