### 2024-07-22 8e3002b
ADD: some more tests
