### 2024-07-22 0de559a
ADD: test for images in mkdocs
