### 2024-07-22 e0acb2d
ADD: additional tests for baseline models (#133)
* ADD: final baseline models test

* ADD: test for images in mkdocs
