+++ tests/unit_tests/models/test_baseline_models.py
@@ -1,6 +1,7 @@
 """Testing dumb - baseline models."""
 
 import numpy as np
+import pytest
 
 from choice_learn.data import ChoiceDataset
 from choice_learn.models.baseline_models import DistribMimickingModel, RandomChoiceModel
@@ -38,6 +39,31 @@ def test_random_choice_model():
     assert (y_pred >= 0).numpy().all()
     assert (y_pred <= 1.0).numpy().all()
 
+    utility = model.compute_batch_utility(
+        shared_features_by_choice=dataset.shared_features_by_choice,
+        items_features_by_choice=dataset.items_features_by_choice,
+        available_items_by_choice=np.ones((4, 3)),
+        choices=dataset.choices,
+    )
+    assert utility.shape == (4, 3)
+
+
+def test_random_choice_model_with_sgd():
+    """Test basic stuff about the RandomChoiceModel with SGD optimizer."""
+    global dataset
+
+    model = RandomChoiceModel(optimizer="SGD")
+    _ = model.fit(choice_dataset=dataset)
+
+    y_pred = model.predict_probas(dataset)
+    assert y_pred.shape == (4, 3)
+    assert (np.sum(y_pred, axis=1) < 1.01).all()
+    assert (np.sum(y_pred, axis=1) > 0.99).all()
+    assert (y_pred >= 0).numpy().all()
+    assert (y_pred <= 1.0).numpy().all()
+
+    assert model.trainable_weights == []
+
 
 def test_mimicking_choice_model():
     """Test basic stuff about the DistribMimickingModel."""
@@ -45,7 +71,6 @@ def test_mimicking_choice_model():
 
     model = DistribMimickingModel()
     _ = model.fit(choice_dataset=dataset)
-
     y_pred = model.predict_probas(dataset)
     assert y_pred.shape == (4, 3)
     assert (np.sum(y_pred, axis=1) < 1.01).all()
@@ -56,3 +81,49 @@ def test_mimicking_choice_model():
     assert (np.abs(y_pred[:, 0] - 0.5) < 0.01).all()
     assert (np.abs(y_pred[:, 1] - 0.25) < 0.01).all()
     assert (np.abs(y_pred[:, 2] - 0.25) < 0.01).all()
+
+    utility = model.compute_batch_utility(
+        shared_features_by_choice=dataset.shared_features_by_choice,
+        items_features_by_choice=dataset.items_features_by_choice,
+        available_items_by_choice=np.ones((4, 3)),
+        choices=dataset.choices,
+    )
+    assert utility.shape == (4, 3)
+    assert (np.abs(utility[0] - utility[3]) < 0.01).all()
+
+
+def test_mimicking_choice_model_with_sgd():
+    """Test basic stuff about the DistribMimickingModel with SGD optimizer."""
+    global dataset
+
+    model = DistribMimickingModel(optimizer="SGD")
+    _ = model.fit(choice_dataset=dataset)
+    y_pred = model.predict_probas(dataset)
+    assert y_pred.shape == (4, 3)
+    assert (np.sum(y_pred, axis=1) < 1.01).all()
+    assert (np.sum(y_pred, axis=1) > 0.99).all()
+    assert (y_pred >= 0).numpy().all()
+    assert (y_pred <= 1.0).numpy().all()
+
+    assert (np.abs(y_pred[:, 0] - 0.5) < 0.01).all()
+    assert (np.abs(y_pred[:, 1] - 0.25) < 0.01).all()
+    assert (np.abs(y_pred[:, 2] - 0.25) < 0.01).all()
+
+    assert (model.trainable_weights.numpy() == np.array([0.5, 0.25, 0.25])).all()
+
+
+def test_catch_not_fitted_issue():
+    """Verify that the model raises an error if not fitted."""
+    global dataset
+
[…]
