### 2024-07-24 #6899 @samuelduchesne
@crvernon, I completed my review. Overall an excellent submission. One comment regarding the `requirements.txt`, I wonder how having strict versions for `numpy`, `pandas` and `tqdm` will play well with other scientific packages a user might want to install in an environment. Perhaps the use of the [compatible release clause](https://packaging.python.org/en/latest/specifications/version-specifiers/#compatible-release) is preferable?
