### 2024-07-24 24359ad
ENH: input_shape is not longer an int in order to work with latest tf_keras distributions
