+++ choice_learn/models/rumnet.py
@@ -7,7 +7,12 @@ from choice_learn.models.base_model import ChoiceModel
 
 
 def create_ff_network(
-    input_shape, depth, width, activation="elu", add_last=False, l2_regularization_coeff=0.0
+    input_shape,
+    depth,
+    width,
+    activation="elu",
+    add_last=False,
+    l2_regularization_coeff=0.0,
 ):
     """Create a simple fully connected (Dense) network.
 
@@ -96,15 +101,15 @@ def recreate_official_nets(
         Features and encoding to utility computation network
     """
     # Products and Customers embeddings nets, quiet symmetrical
-    products_input = tf.keras.layers.Input(shape=(num_products_features))
-    customer_input = tf.keras.layers.Input(shape=(num_customer_features))
+    products_input = tf.keras.layers.Input(shape=(num_products_features,))
+    customer_input = tf.keras.layers.Input(shape=(num_customer_features,))
     x_embeddings = []
     z_embeddings = []
 
     # Creating independant nets for each heterogeneity
     for _ in range(x_eps):
         x_embedding = create_ff_network(
-            input_shape=num_products_features,
+            input_shape=(num_products_features,),
             depth=x_depth,
             width=x_width,
             l2_regularization_coeff=l2_regularization_coeff,
@@ -114,7 +119,7 @@ def recreate_official_nets(
     # Creating independant nets for each heterogeneity
     for _ in range(z_eps):
         z_embedding = create_ff_network(
-            input_shape=num_customer_features,
+            input_shape=(num_customer_features,),
             depth=z_depth,
             width=z_width,
             l2_regularization_coeff=l2_regularization_coeff,
@@ -122,13 +127,17 @@ def recreate_official_nets(
 
         z_embeddings.append(z_embedding)
 
-    x_net = tf.keras.Model(inputs=products_input, outputs=x_embeddings, name="X_embedding")
-    z_net = tf.keras.Model(inputs=customer_input, outputs=z_embeddings, name="Z_embedding")
+    x_net = tf.keras.Model(
+        inputs=products_input, outputs=x_embeddings, name="X_embedding"
+    )
+    z_net = tf.keras.Model(
+        inputs=customer_input, outputs=z_embeddings, name="Z_embedding"
+    )
 
     # Utility network
     u_net = create_ff_network(
         input_shape=(
-            x_width + z_width + num_products_features + num_customer_features
+            x_width + z_width + num_products_features + num_customer_features,
         ),  # Input shape from previous nets
         width=width_u,
         depth=depth_u,
@@ -587,7 +596,9 @@ class PaperRUMnet(ChoiceModel):
         )
 
         # Storing weights for back-propagation
-        self.trainable_weights = self.x_model.weights + self.z_model.weights + self.u_model.weights
+        self.trainable_weights = (
+            self.x_model.weights + self.z_model.weights + self.u_model.weights
+        )
         self.loss = tf_ops.CustomCategoricalCrossEntropy(
             from_logits=False,
             label_smoothing=self.label_smoothing,
@@ -640,7 +651,10 @@ class PaperRUMnet(ChoiceModel):
             )
         if isinstance(items_features_by_choice, tuple):
             items_features_by_choice = tf.concat(
-                [tf.cast(items_feature, tf.float32) for items_feature in items_features_by_choice],
+                [
+                    tf.cast(items_feature, tf.float32)
+                    for items_feature in items_features_by_choice
+                ],
                 axis=-1,
             )
 
@@ -662,7 +676,12 @@ class PaperRUMnet(ChoiceModel):
             for _x in x_embeddings:
                 for _z in z_embeddings:
                     _u = tf.keras.layers.Concatenate()(
[…]
