### 2024-07-24 32d63f4
ENH: requirements allow larger package versions
