+++ requirements.txt
@@ -1,6 +1,6 @@
--e .
-numpy==1.24.3
-pandas==1.5.3
-tensorflow==2.14.0
-tensorflow_probability==0.22.1
-tqdm==4.65.0
+numpy>=1.24.3
+pandas~=1.5
+tensorflow~=2.14
+tensorflow_probability~=0.22
+tf_keras~=2.17
+tqdm~=4.0
