### 2024-07-24 #6899 @crvernon
Thanks @samuelduchesne!  @VincentAuriau is there a reason for pinning the package versions?  What are your thoughts on making these more flexible to greatly increase compatibility with other packages?
