+++ .github/workflows/pr_ci.yaml
@@ -34,3 +34,4 @@ jobs:
         with:
           pytest-coverage-path: ./pytest-coverage.txt
           junitxml-path: ./pytest.xml
+          report-only-changed-files: true
