### 2024-07-24 fd037cc
ENH: naming trainable_weights -> _trainable_weights
