+++ choice_learn/models/base_model.py
@@ -126,14 +126,17 @@ class ChoiceModel(object):
             self.regularization = None
 
     @property
-    @abstractmethod
     def trainable_weights(self):
         """Trainable weights need to be specified in children classes.
 
         Basically it determines which weights need to be optimized during training.
         MUST be a list
         """
-        return
+        raise NotImplementedError(
+            """Trainable_weights must be specified in children classes,
+              when you inherit from ChoiceModel.
+            See custom models documentation for more details and examples."""
+        )
 
     @abstractmethod
     def compute_batch_utility(
