### 2024-07-24 56eb01f
FIX: trainable_weights of DistribMimicking model
