### 2024-07-25 118b461
ENH: requirements.txt allow for broader versions of dependencies (#134)
* ENH: input_shape is not longer an int in order to work with latest tf_keras distributions

* ENH: requirements allow larger package versions
