### 2024-07-25 4c6a9d8
ADD: forgotten changes for _trainable_weights
