+++ choice_learn/models/simple_mnl.py
@@ -309,7 +309,7 @@ class SimpleMNL(ChoiceModel):
                 for _w in self.trainable_weights:
                     mw.append(w[index : index + _w.shape[0]])
                     index += _w.shape[0]
-                model.weights = mw
+                model._trainable_weights = mw
                 for batch in choice_dataset.iter_batch(batch_size=-1):
                     utilities = model.compute_batch_utility(*batch)
                     probabilities = tf.nn.softmax(utilities, axis=-1)
@@ -341,7 +341,7 @@ class SimpleMNL(ChoiceModel):
         if hasattr(self, "report"):
             clone.report = self.report
         if hasattr(self, "trainable_weights"):
-            clone.weights = self.trainable_weights
+            clone._trainable_weights = self.trainable_weights
         if hasattr(self, "indexes"):
             clone.indexes = self.indexes
         if hasattr(self, "intercept"):
