+++ .github/workflows/ci.yaml
@@ -34,4 +34,5 @@ jobs:
         with:
           pytest-coverage-path: ./pytest-coverage.txt
           junitxml-path: ./pytest.xml
-          issue-number: 101
+          issue-number: '101'
+          unique-id-for-comment: ${{ matrix.python-version }}
