+++ .github/workflows/pr_ci.yaml
@@ -35,3 +35,4 @@ jobs:
           pytest-coverage-path: ./pytest-coverage.txt
           junitxml-path: ./pytest.xml
           report-only-changed-files: true
+          unique-id-for-comment: ${{ matrix.python-version }}
