+++ .github/workflows/ci.yaml
@@ -36,3 +36,4 @@ jobs:
           junitxml-path: ./pytest.xml
           issue-number: '101'
           unique-id-for-comment: ${{ matrix.python-version }}
+          title: 'Coverage Report for Python ${{ matrix.python-version }}'
