+++ .github/workflows/pr_ci.yaml
@@ -36,3 +36,4 @@ jobs:
           junitxml-path: ./pytest.xml
           report-only-changed-files: true
           unique-id-for-comment: ${{ matrix.python-version }}
+          title: 'Coverage Report for Python ${{ matrix.python-version }}'
