+++ .github/workflows/pr_ci.yaml
@@ -37,3 +37,9 @@ jobs:
           report-only-changed-files: true
           unique-id-for-comment: ${{ matrix.python-version }}
           title: 'Coverage Report for Python ${{ matrix.python-version }}'
+      
+      name: Test & echo
+      run: |
+        echo ${{ matrix.python-version }}
+        echo "Test"
+        echo ${{ steps.pr_number.outputs.content }}
