### 2024-07-25 efff23c
ENH: ChoiceModel.trainable_weights as @property (#116)
* FIX: ArrayStorage when directly instantiated from ndarray

* ENH: better handling when features / availabilities are indexed with FeaturesStorage

* ADD: ArrayStorage indexer & catching KeyErrors for custom messaging

* ENH: ArrayStorage follows common structure with Indexers

* ENH: harmonize Error raised with missing key

* ADD: KeyError test for all Storages

* ENH: FeaturesStorage IDs checked automatically

* ENH: Way faster FeaturesStorage instantiation in some edge cases

* ENH: trainable_weights as abstractmethod & property

* FIX
[…]
