+++ notebooks/models/custom_modelling.md
@@ -1,1 +1,1 @@
-Custom modelling with Choice-Learn is part of the introductive tutorial and is detailed [here](../introduction/3_model_clogit.ipynb).
+Custom modelling with Choice-Learn is part of the introductive tutorial and is detailed [here](../introduction/4_model_customization.ipynb).
