+++ .github/workflows/pr_ci.yaml
@@ -44,3 +44,12 @@ jobs:
           echo "Test"
           echo ${{ github.event.number }}
           echo "Test"
+
+      - name: Pytest coverage comment
+        uses: MishaKav/pytest-coverage-comment@main
+        with:
+          pytest-coverage-path: ./pytest-coverage.txt
+          junitxml-path: ./pytest.xml
+          issue-number: '101'
+          unique-id-for-comment: ${{ matrix.python-version }}
+          title: 'Coverage Report for Python ${{ matrix.python-version }}'
