### 2024-07-25 ffa3bda
Change version to v0.0.6 (#138)
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
