+++ pyproject.toml
@@ -55,7 +55,7 @@ pythonpath = ["./"]
 
 [tool.poetry]
 name = "choice-learn"
-version = "0.0.5"
+version = "0.0.6"
 description = "Large-scale choice modeling through the lens of machine learning."
 authors = [
   "artefactory <vincent.auriau@artefact.com>",
