### 2024-07-31 69ee06c
ENH: broader poetry dependencies
