### 2024-07-31 fe0b1f6
ADD: broader pip requirements
