### 2024-07-31 87aaf10
ADD: broader pip requirements
