### 2024-07-31 be89983
ENH: broader poetry/pip dependencies (#139)
