+++ requirements.txt
@@ -1,6 +1,6 @@
-numpy>=1.24.3
-pandas~=1.5
-tensorflow~=2.14
+numpy>=1.24
+pandas>=1.5
+tensorflow>=2.14, <2.17
 tensorflow_probability~=0.22
-tf_keras~=2.17
-tqdm~=4.0
+tf_keras<3
+tqdm>=4.0
