+++ pyproject.toml
@@ -45,13 +45,12 @@ quote-style = "double"
 [tool.ruff.isort]
 known-first-party = ["choice_learn", "config", "tests"]
 
-[tool.ruff.lint]
-
 [tool.bandit.assert_used]
 exclude_dirs = ["tests/"]
 
 [tool.pytest.ini_options]
 pythonpath = ["./"]
+testpaths = ["tests/"]
 
 [tool.poetry]
 name = "choice-learn"
@@ -76,33 +75,33 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.9.0"
-numpy = "^1.24.3"
-pandas = "^1.5.3"
-tensorflow = "^2.14.0"
-tensorflow-probability = "^0.22.1"
-tqdm = "^4.0.0"
-ortools = { version = "^9.6", optional = true }
-gurobipy = { version = "^11.0", optional = true }
+python = "~3.9"
+numpy = ">=1.24"
+pandas = ">= 1.5.3"
+tensorflow = "~2.14, <2.17.0"
+tensorflow-probability = "~0.22"
+tqdm = ">=4.0.0"
+ortools = { version = "~9.6", optional = true }
+gurobipy = { version = "~11.0", optional = true }
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^7.3.2"
-pre-commit = "^3.3"
-ipykernel = "^6.9"
-nbstripout = "^0.7"
-ruff = "^0.1.2"
-pytest-cov = "^4.1"
-codespell = "^2.2"
-python-markdown-math = "^0.8"
+pytest = ">=7.3.2"
+pre-commit = ">=3.3"
+ipykernel = ">=6.9"
+nbstripout = ">=0.7"
+ruff = ">=0.1.2"
+pytest-cov = ">=4.1"
+codespell = ">=2.2"
+python-markdown-math = ">=0.8"
 
 [tool.poetry.group.docs.dependencies]
-mkdocs = "^1.5"
-mkdocs-material = {extras = ["pygments"], version = "^9.5.3"}
-mkdocs-nbconvert = "^0.2.1"
-mkdocstrings-python = "^1.7.5"
-python-markdown-math = "^0.8"
-bandit = "^1.7.5"
-nbstripout = "^0.7"
+mkdocs = ">=1.5"
+mkdocs-material = {extras = ["pygments"], version = ">=9.5.3"}
+mkdocs-nbconvert = ">=0.2.1"
+mkdocstrings-python = ">=1.7.5"
+python-markdown-math = ">=0.8"
+bandit = ">=1.7.5"
+nbstripout = ">=0.7"
 
 [build-system]
 requires = ["poetry-core"]
