### 2024-07-31 5f297bc
ADD: Python 3.12 in CI
