### 2024-07-31 97a5721
Change version to v0.0.7 (#140)
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
