+++ .github/workflows/pr_ci.yaml
@@ -46,7 +46,7 @@ jobs:
           echo "Test"
 
       - name: Pytest coverage comment
-        uses: MishaKav/pytest-coverage-comment@main
+        uses: VincentAuriau/pytest-coverage-comment@main
         with:
           pytest-coverage-path: ./pytest-coverage.txt
           junitxml-path: ./pytest.xml
