+++ .github/workflows/ci.yaml
@@ -3,6 +3,7 @@ on:
   push:
     branches:
       - main
+      - enh/cov
 
 jobs:
   CI:
@@ -30,7 +31,7 @@ jobs:
           pytest --junitxml=pytest.xml --cov-report=term-missing:skip-covered --cov=choice_learn tests/ | tee pytest-coverage.txt
 
       - name: Pytest coverage comment
-        uses: MishaKav/pytest-coverage-comment@main
+        uses: VincentAuriau/pytest-coverage-comment@main
         with:
           pytest-coverage-path: ./pytest-coverage.txt
           junitxml-path: ./pytest.xml
