### 2024-08-01 #6899 @VincentAuriau
Hello, coming back to you on this dependencies subject, the basics requirements are now:

- numpy>=1.24
- pandas>=1.5
- tensorflow>=2.14, <2.17
- tensorflow_probability~=0.22
- tf_keras<3
- tqdm>=4.0

Everything seems to be working fine with it !
