+++ requirements.txt
@@ -2,5 +2,5 @@ numpy>=1.24
 pandas>=1.5
 tensorflow>=2.14, <2.17
 tensorflow_probability~=0.22
-tf_keras<3
+tf_keras<3; tensorflow>=2.16
 tqdm>=4.0
