+++ pyproject.toml
@@ -80,7 +80,8 @@ numpy = ">=1.24"
 pandas = ">= 1.5.3"
 tensorflow = ">=2.14, <2.17.0"
 tensorflow-probability = "~0.22"
-tqdm = ">=4.0.0"
+tqdm = ">=4.0.0
+tf-keras = { version = "<3>", tensorflow = ">2.16" }
 ortools = { version = ">=9.6", optional = true }
 gurobipy = { version = ">=11.0", optional = true }
 
