+++ requirements.txt
@@ -1,6 +1,5 @@
 numpy>=1.24
 pandas>=1.5
 tensorflow>=2.14, <2.17
-tensorflow_probability~=0.22
-tf_keras<3; tensorflow>=2.16
+tensorflow_probability>=0.22
 tqdm>=4.0
