+++ pyproject.toml
@@ -81,6 +81,7 @@ pandas = ">= 1.5.3"
 tensorflow = ">=2.14, <2.17.0"
 tensorflow-probability = ">=0.22"
 tqdm = ">=4.0.0"
+tf_keras = "<3"
 ortools = { version = ">=9.6", optional = true }
 gurobipy = { version = ">=11.0", optional = true }
 
