+++ requirements.txt
@@ -3,4 +3,4 @@ pandas~=1.5
 tensorflow~=2.14
 tensorflow_probability~=0.22
 tf_keras~=2.17
-tqdm~=4.0
+tqdm>=4.0
