+++ choice_learn/__init__.py
@@ -1,4 +1,4 @@
 """Choice-Learn library for choice modeling in Python."""
 
-__version__ = "0.0.7"
+__version__ = "0.0.8"
 __author__ = "Vincent Auriau"
