+++ .github/workflows/ci.yaml
@@ -3,7 +3,6 @@ on:
   push:
     branches:
       - main
-      - enh/cov
 
 jobs:
   CI:
