### 2024-08-02 b7b3ef7
Enh/cov (#137)
* Automatic Tracking of coverage % in Issue 101
* FIX: softmax with exit choice tests
* ADD: softmax with exit choice tests
