### 2024-08-02 #6899 @samuelduchesne
I know numpy released v2 recently. Have you tested using that version?
