### 2024-08-02 #6899 @VincentAuriau
I have just made a few tests, it looks like Tensorflow requires numpy < 2. I will further look into it !
