### 2024-08-02 #6899 @samuelduchesne
In that case, `pip` should resolve the `numpy` version properly (<2). In any case, your package is not limiting that choice with `numpy>=1.24`. In other words, this looks good!
