### 2024-08-06 artefactory/choice-learn#145 @tmigot
[ADD] Comments on Experiments and examples JOSS paper
## 🚀 Feature Request

Suggestions regarding the *Experiments and examples* section of the paper submitted to JOSS :
- [x] The (Ben Hamner, Friedman, 2013) citation seems to be very Kaggle specific (for instance I didn't find a correspondance on Google Scholar), so could you add on the reference to add the link to the data. Kaggle website suggests:
```
Adam, Ben Hamner, Dan Friedman, SSA_Expedia. (2013). Personalize Expedia Hotel Searches - ICDM 2013. Kaggle. https://kaggle.com/competitions/expedia-personalized-sort
```
In general having links for each reference is a good practice in such journal.
- [x] Is there a notebook somewhere on this repo that shows the code used to generate this benchmark? If so, I would add the link to it in a footnote of the paper or at least a comment that it is in the repository. There is a lot of content here (congrat's!), but it would really help the reader to find its way.
- [x] The "customize choice models" is indeed a nice feature, however I'm not sure it is necessary to show code regarding this in the paper. It is probably sufficient to just mention it and link to the documentation of the package.

## 📎 Additional context

This comment are here to ensure data is accessible and reproducibility.

Connected to https://github.com/openjournals/joss-reviews/issues/6899
