### 2024-08-09 artefactory/choice-learn#145 @VincentAuriau
Hello @tmigot,
Thank you for your review !

- [Here](https://github.com/artefactory/choice-learn/tree/joss/docs/paper) are two notebooks for the different experiments (benchmark + custom model)
- I will modify the paper following your recommendations.
