### 2024-08-09 artefactory/choice-learn#146 @tmigot
[ADD] Comments on JOSS paper
## 🚀 Feature Request

Suggestions regarding JOSS paper :
- [x] p.3 L.45: I think it should be made clearer that the `FeaturesStorage` structure comes from this packages and not numpy
- [x] p.3 L.51: L-BFGS is not a second-order algorithm as it doesn't use any second-order information on the problem. It is true though that it uses first-order information to build an approximation of the second-order.
- [x] The paper is quite long for JOSS format, and it would be great (also optional IMO), one suggestion would be to remove the *Streamlined signature* section. In general, the paper should focus on the main points.
## 📎 Additional context

Connected to https://github.com/openjournals/joss-reviews/issues/6899
