### 2024-08-09 #6899 @tmigot
@crvernon I am done with my review. This is great work from @VincentAuriau and his co-authors. I think the package is well-organized and will be very useful.
I opened two issues directly on the package with minor issues, and will happily finish ticking my list once they are completed.
