+++ docs/paper/paper.bib
@@ -137,6 +137,7 @@ doi = {10.1007/0-306-48332-7_250}
   journal={ICDM},
       year={2013},
       howpublished={\url{https://www.kaggle.com/c/expedia-personalized-sort} (last checked on 2024-06-07)},
+      url={\url{https://www.kaggle.com/c/expedia-personalized-sort}},
 }
 
 @article{AouadMarket:2023,
@@ -192,13 +193,6 @@ doi = {10.1016/j.dam.2012.03.003},
   year={2011}
 }
 
-@misc{Helveston:2023,
-title={Convert data from wide to long format},
-howpublished={\url{https://xlogit.readthedocs.io/en/latest/notebooks/convert_data_wide_to_long.html} (last checked on 2024-06-07)},
-author={Forsythe, C. and Helveston, J},
-  year={2023},
-  urldate     = {2024-06-07},
-}
 @misc{Chollet:2015,
   title={Keras},
   author={{Chollet et al.}},
