+++ docs/paper/paper.bib
@@ -137,7 +137,7 @@ doi = {10.1007/0-306-48332-7_250}
   journal={ICDM},
       year={2013},
       howpublished={\url{https://www.kaggle.com/c/expedia-personalized-sort} (last checked on 2024-06-07)},
-      url={\url{https://www.kaggle.com/c/expedia-personalized-sort}},
+      url={https://www.kaggle.com/c/expedia-personalized-sort},
 }
 
 @article{AouadMarket:2023,
