+++ docs/paper/paper.bib
@@ -132,12 +132,11 @@ doi = {10.1007/0-306-48332-7_250}
 }
 
 @article{Expedia:2013,
-      title={Personalize Expedia Hotel Searches},
-      author={Ben Hamner, Adam and Friedman, Dan},
-  journal={ICDM},
-      year={2013},
-      howpublished={\url{https://www.kaggle.com/c/expedia-personalized-sort} (last checked on 2024-06-07)},
-      url={https://www.kaggle.com/c/expedia-personalized-sort},
+  title={Personalize Expedia Hotel Searches - ICDM 2013},
+  author={Ben Hamner, Adam and Friedman, Dan and SSA_Expedia},
+  journal={Kaggle},
+  year={2013},
+  url={https://www.kaggle.com/c/expedia-personalized-sort},
 }
 
 @article{AouadMarket:2023,
