+++ docs/paper/paper.bib
@@ -196,7 +196,7 @@ doi = {10.1016/j.dam.2012.03.003},
   title={Keras},
   author={{Chollet et al.}},
   year={2015},
-  howpublished={\url{https://keras.io} (last checked on 2024-06-07)},
+  howpublished={\url{https://keras.io}},
 }
 
 @article{McFadden:2000,
@@ -215,7 +215,7 @@ doi = {10.1016/j.dam.2012.03.003},
   author = {{Gurobi Optimization, LLC}},
   title = {{Gurobi Optimizer Reference Manual}},
   year = 2023,
-  howpublished = {\url{https://www.gurobi.com} (last checked on 2024-06-07)}
+  howpublished = {\url{https://www.gurobi.com}}
 }
 
 @software{ORTools:2024,
@@ -224,5 +224,5 @@ doi = {10.1016/j.dam.2012.03.003},
   author = {Laurent Perron and Vincent Furnon},
   organization = {Google},
   date = { 2024-03-07 },
-  howpublished = {\url{https://developers.google.com/optimization/} (last checked on 2024-06-07)}
+  howpublished = {\url{https://developers.google.com/optimization/}}
 }
