### 2024-08-11 ca43710
Enhance Paper (#147)
* ENH: modifs from review

* FIX: url

* FIX: citation

* FIX: citation

* ENH: customized model

* ENH: add .
