+++ docs/paper/paper.bib
@@ -132,11 +132,11 @@ doi = {10.1007/0-306-48332-7_250}
 }
 
 @article{Expedia:2013,
-      title={Personalize Expedia Hotel Searches},
-      author={Ben Hamner, Adam and Friedman, Dan},
-  journal={ICDM},
-      year={2013},
-      howpublished={\url{https://www.kaggle.com/c/expedia-personalized-sort} (last checked on 2024-06-07)},
+  title={Personalize Expedia Hotel Searches - ICDM 2013},
+  author={Ben Hamner, Adam and Friedman, Dan and SSA_Expedia},
+  journal={Kaggle},
+  year={2013},
+  url={https://www.kaggle.com/c/expedia-personalized-sort},
 }
 
 @article{AouadMarket:2023,
@@ -192,18 +192,11 @@ doi = {10.1016/j.dam.2012.03.003},
   year={2011}
 }
 
-@misc{Helveston:2023,
-title={Convert data from wide to long format},
-howpublished={\url{https://xlogit.readthedocs.io/en/latest/notebooks/convert_data_wide_to_long.html} (last checked on 2024-06-07)},
-author={Forsythe, C. and Helveston, J},
-  year={2023},
-  urldate     = {2024-06-07},
-}
 @misc{Chollet:2015,
   title={Keras},
   author={{Chollet et al.}},
   year={2015},
-  howpublished={\url{https://keras.io} (last checked on 2024-06-07)},
+  howpublished={\url{https://keras.io}},
 }
 
 @article{McFadden:2000,
@@ -222,7 +215,7 @@ author={Forsythe, C. and Helveston, J},
   author = {{Gurobi Optimization, LLC}},
   title = {{Gurobi Optimizer Reference Manual}},
   year = 2023,
-  howpublished = {\url{https://www.gurobi.com} (last checked on 2024-06-07)}
+  howpublished = {\url{https://www.gurobi.com}}
 }
 
 @software{ORTools:2024,
@@ -231,5 +224,5 @@ author={Forsythe, C. and Helveston, J},
   author = {Laurent Perron and Vincent Furnon},
   organization = {Google},
   date = { 2024-03-07 },
-  howpublished = {\url{https://developers.google.com/optimization/} (last checked on 2024-06-07)}
+  howpublished = {\url{https://developers.google.com/optimization/}}
 }
