### 2024-08-11 artefactory/choice-learn#145 @VincentAuriau
Hello,
I fixed the the citation and removed the example, while adding a sentence earlier in the paper to link the documentation !
