### 2024-08-11 artefactory/choice-learn#146 @VincentAuriau
- I have added '''The package introduces the *FeaturesStorage* object''' which makes clearer that it's not from NumPy
- I changed second-order to quasi-Newton
- I removed the paragraph while keeping a word about this streamlined objective in the introduction. It has taken some discussions and time to finally reach the current version that's why I like having a word about it. What do you think about the current version ?

Thank you for your recommendations that help improving the paper !
