### 2024-08-11 artefactory/choice-learn#145 @tmigot
thanks @VincentAuriau
