### 2024-08-19 598e54c
ADD: tf_ops tests
