### 2024-08-19 7ecf53f
ADD: Indexer tests
