### 2024-08-19 ff0b7ed
ENH: Ruff
