+++ tests/unit_tests/data/test_indexer.py
@@ -1,8 +1,8 @@
 """Test ChoiceDataset and Storage Indexers."""
 
-import numpy as np
 
 from choice_learn.data.indexer import Indexer
+
 
 def test_dummy_indexer():
     """Dummy test for the base class Indexer."""
