+++ tests/unit_tests/data/test_choice_dataset.py
@@ -1,4 +1,5 @@
 """Unit testing for class ChoiceDataset."""
+
 import pandas as pd
 import pytest
 
