### 2024-08-19 257d4f4
ADD: CE label smoothing test
