### 2024-08-19 f43405d
FIX: OneHotStorage with slice & ADD: corresponding test
