### 2024-08-19 edb7247
FIX: some indexing & ADD: corresponding tests
