### 2024-08-20 cd1184b
FIX: get_full_datset with FS & ADD: finalized indexer coverage
