### 2024-08-20 9b7d704
ADD: various CD instantiation tests
