### 2024-08-20 acf7a18
ADD: CD instantiation tests
