### 2024-08-20 25a8da0
ADD: tests & some fixes
