### 2024-08-20 f2ae4b4
ADD: some Storage tests (#149)
* ADD: some Storage tests

* ADD: tests & some fixes
