+++ tests/unit_tests/data/test_indexer.py
@@ -1,10 +1,11 @@
 """Test ChoiceDataset and Storage Indexers."""
 
+import pytest
+
 from choice_learn.data.indexer import Indexer
 
 
 def test_dummy_indexer():
     """Dummy test for the base class Indexer."""
-    indexer = Indexer(list)
-    indexer[0]
-    assert True
+    with pytest.raises(TypeError):
+        Indexer(list)
