### 2024-08-21 1394420
ADD: some more tests for ChoiceDataset
