+++ tests/unit_tests/data/test_choice_dataset.py
@@ -49,12 +49,75 @@ items_features_by_choice = [
 def test_raise_error():
     """Test various errors in instantiation."""
     with pytest.raises(ValueError):
+        # Choices is None
         ChoiceDataset(
             shared_features_by_choice=shared_features_by_choice,
             items_features_by_choice=items_features_by_choice,
             available_items_by_choice=available_items_by_choice,
             choices=None,
         )
+    with pytest.raises(ValueError):
+        # wrong items features names shape
+        ChoiceDataset(
+            shared_features_by_choice=shared_features_by_choice,
+            items_features_by_choice=(items_features_by_choice, items_features_by_choice),
+            items_features_by_choice_names=(["ab", "bc"], ["cd"]),
+            available_items_by_choice=available_items_by_choice,
+            choices=choices,
+        )
+    with pytest.raises(ValueError):
+        # wrong shared features names shape
+        ChoiceDataset(
+            shared_features_by_choice=(shared_features_by_choice, shared_features_by_choice),
+            shared_features_by_choice_names=(["a", "b"], ["c", "d", "e"]),
+            items_features_by_choice=items_features_by_choice,
+            available_items_by_choice=available_items_by_choice,
+            choices=choices,
+        )
+
+    with pytest.raises(ValueError):
+        # wrong nb of items in items features shape
+        ChoiceDataset(
+            shared_features_by_choice=shared_features_by_choice,
+            items_features_by_choice=(
+                items_features_by_choice,
+                np.array(
+                    [
+                        [
+                            [100, 0, 2],
+                            [140, 0, 2],
+                            [200, 0, 2],
+                            [200, 0, 2],
+                        ],
+                        [
+                            [100, 0, 2],
+                            [120, 1, 2],
+                            [200, 0, 2],
+                            [200, 0, 2],
+                        ],
+                        [
+                            [0, 0, 2],
+                            [120, 1, 2],
+                            [180, 1, 2],
+                            [180, 1, 2],
+                        ],
+                    ]
+                ),
+            ),
+            available_items_by_choice=available_items_by_choice,
+            choices=[0, 2, 1],
+        )
+
+    # wrong nb of shared features names
+    with pytest.raises(ValueError):
+        ChoiceDataset(
+            shared_features_by_choice=shared_features_by_choice,
+            items_features_by_choice=items_features_by_choice,
+            available_items_by_choice=np.array([[1, 1], [0, 1], [1, 0]]),
+            choices=[0, 2, 1],
+        )
+
+    # wrong nb of shared features names
     with pytest.raises(ValueError):
         ChoiceDataset(
             shared_features_by_choice=shared_features_by_choice,
@@ -94,6 +157,37 @@ def test_raise_error():
             items_features_by_choice_names=(["ab", "bc"], ["abc"]),
             available_items_by_choice=available_items_by_choice,
             choices=[0, 2, 1],
+        )
+    with pytest.raises(ValueError):
+        ChoiceDataset(
+            shared_features_by_choice=shared_features_by_choice,
+            items_features_by_choice=(items_features_by_choice, [0, 1, 0]),
+            items_features_by_choice_names=(["ab", "bc"], ["abc", "def"]),
[…]
