+++ pyproject.toml
@@ -96,10 +96,11 @@ codespell = ">=2.2"
 python-markdown-math = ">=0.8"
 
 [tool.poetry.group.docs.dependencies]
+griffe = "<0.49"
 mkdocs = ">=1.5"
 mkdocs-material = {extras = ["pygments"], version = ">=9.5.3"}
 mkdocs-nbconvert = ">=0.2.1"
-mkdocstrings-python = ">=1.10.8"
+mkdocstrings-python = ">=1.7.5"
 python-markdown-math = ">=0.8"
 bandit = ">=1.7.5"
 nbstripout = ">=0.7"
