### 2024-08-21 779177d
ADD: some more tests (#150)
* ADD: some more tests for ChoiceDataset

* FIX: mkdocstring dependencies ?

* FIX: linting from online conflict resolution...
