### 2024-08-21 60ef16b
ADD: griffe as dev requirement
