+++ requirements-developer.txt
@@ -1,4 +1,5 @@
 -r requirements.txt
+griffe<=0.48
 ruff==0.1.2
 pre-commit==3.3.3
 pytest==7.3.2
