### 2024-08-22 39fb303
ENH: get_latent weight used in .fit
