### 2024-08-23 c59ae0e
ADD: Latent Class with gradient descent
