+++ choice_learn/models/latent_class_base_model.py
@@ -82,7 +82,7 @@ class BaseLatentClassModel(object):
     def instantiate(self, **kwargs):
         """Instantiate the model."""
         init_logit = tf.Variable(
-            tf.random_normal_initializer(0.0, 0.02, seed=42)(shape=(self.n_latent_classes - 1,)),
+            tf.random_normal_initializer(0.0, 0.08, seed=42)(shape=(self.n_latent_classes - 1,)),
             name="Latent-Logits",
         )
         self.latent_logits = init_logit
@@ -240,7 +240,7 @@ class BaseLatentClassModel(object):
                     choice_dataset=choice_dataset, sample_weight=sample_weight, verbose=verbose
                 )
 
-            return self._fit_normal(
+            return self._fit_with_gd(
                 choice_dataset=choice_dataset, sample_weight=sample_weight, verbose=verbose
             )
 
@@ -457,10 +457,267 @@ class BaseLatentClassModel(object):
             print("---------------------------------------------------------------")
             print("Number of iterations:", results[2].numpy())
             print("Algorithm converged before reaching max iterations:", results[0].numpy())
-        return func.history
+        return func.history, results
 
-    def _gd_train_step(self, choice_dataset, sample_weight=None):
-        pass
+    @tf.function
+    def train_step(
+        self,
+        shared_features_by_choice,
+        items_features_by_choice,
+        available_items_by_choice,
+        choices,
+        sample_weight=None,
+    ):
+        """Represent one training step (= one gradient descent step) of the model.
+
+        Parameters
+        ----------
+        shared_features_by_choice : tuple of np.ndarray (choices_features)
+            a batch of shared features
+            Shape must be (n_choices, n_shared_features)
+        items_features_by_choice : tuple of np.ndarray (choices_items_features)
+            a batch of items features
+            Shape must be (n_choices, n_items_features)
+        available_items_by_choice : np.ndarray
+            A batch of items availabilities
+            Shape must be (n_choices, n_items)
+        choices_batch : np.ndarray
+            Choices
+            Shape must be (n_choices, )
+        sample_weight : np.ndarray, optional
+            List samples weights to apply during the gradient descent to the batch elements,
+            by default None
+
+        Returns
+        -------
+        tf.Tensor
+            Value of NegativeLogLikelihood loss for the batch
+        """
+        with tf.GradientTape() as tape:
+            utilities = self.compute_batch_utility(
+                shared_features_by_choice=shared_features_by_choice,
+                items_features_by_choice=items_features_by_choice,
+                available_items_by_choice=available_items_by_choice,
+                choices=choices,
+            )
+
+            probabilities = tf_ops.softmax_with_availabilities(
+                items_logit_by_choice=utilities,
+                available_items_by_choice=available_items_by_choice,
+                normalize_exit=self.add_exit_choice,
+                axis=-1,
+            )
+            # Negative Log-Likelihood
+            neg_loglikelihood = self.loss(
+                y_pred=probabilities,
+                y_true=tf.one_hot(choices, depth=probabilities.shape[1]),
+                sample_weight=sample_weight,
+            )
+            if self.regularization is not None:
+                regularization = tf.reduce_sum(
+                    [self.regularizer(w) for w in self.trainable_weights]
+                )
+                neg_loglikelihood += regularization
+
+        grads = tape.gradient(neg_loglikelihood, self.trainable_weights)
+        self.optimizer.apply_gradients(zip(grads, self.trainable_weights))
+        return neg_loglikelihood
+    
+
+    def _fit_with_gd(
+        self,
+        choice_dataset,
+        sample_weight=None,
+        val_dataset=None,
+        verbose=0,
+    ):
+        """Train the model with a ChoiceDataset.
+
+        Parameters
+        ----------
+        choice_dataset : ChoiceDataset
[…]
