### 2024-08-23 b8a0915
ADD: running gradient descent for Latent Class
