+++ choice_learn/models/latent_class_base_model.py
@@ -2,6 +2,7 @@
 
 import numpy as np
 import tensorflow as tf
+import time
 import tqdm
 
 import choice_learn.tf_ops as tf_ops
@@ -17,6 +18,7 @@ class BaseLatentClassModel(object):
         model_parameters,
         fit_method,
         epochs,
+        batch_size=128,
         optimizer=None,
         add_exit_choice=False,
         tolerance=1e-6,
@@ -64,6 +66,7 @@ class BaseLatentClassModel(object):
         self.tolerance = tolerance
         self.optimizer = optimizer
         self.lr = lr
+        self.batch_size = batch_size
 
         self.loss = tf_ops.CustomCategoricalCrossEntropy(
             from_logits=False, label_smoothing=0.0
@@ -79,6 +82,19 @@ class BaseLatentClassModel(object):
         )
         self.instantiated = False
 
+    @property
+    def trainable_weights(self):
+        """Returns trainable weights.
+
+        Returns
+        -------
+        list
+           list of trainable weights."""
+        weights = [self.latent_logits]
+        for model in self.models:
+            weights += model.trainable_weights
+        return weights
+    
     def instantiate(self, **kwargs):
         """Instantiate the model."""
         init_logit = tf.Variable(
@@ -239,7 +255,16 @@ class BaseLatentClassModel(object):
                 return self._fit_with_lbfgs(
                     choice_dataset=choice_dataset, sample_weight=sample_weight, verbose=verbose
                 )
-
+            if self.optimizer.lower() == "adam":
+                self.optimizer = tf.keras.optimizers.Adam(self.lr)
+            elif self.optimizer.lower() == "sgd":
+                self.optimizer = tf.keras.optimizers.SGD(self.lr)
+            elif self.optimizer.lower() == "adamax":
+                self.optimizer = tf.keras.optimizers.Adamax(self.lr)
+            else:
+                print(f"Optimizer {self.optimizer} not implemnted, switching for default Adam")
+                self.optimizer = tf.keras.optimizers.Adam(self.lr)
+                
             return self._fit_with_gd(
                 choice_dataset=choice_dataset, sample_weight=sample_weight, verbose=verbose
             )
@@ -459,7 +484,7 @@ class BaseLatentClassModel(object):
             print("Algorithm converged before reaching max iterations:", results[0].numpy())
         return func.history, results
 
-    @tf.function
+    # @tf.function
     def train_step(
         self,
         shared_features_by_choice,
@@ -501,23 +526,30 @@ class BaseLatentClassModel(object):
                 choices=choices,
             )
 
-            probabilities = tf_ops.softmax_with_availabilities(
-                items_logit_by_choice=utilities,
-                available_items_by_choice=available_items_by_choice,
-                normalize_exit=self.add_exit_choice,
-                axis=-1,
-            )
+            latent_probabilities = self.get_latent_classes_weights()
+            # Compute probabilities from utilities & availabilties
+            probabilities = []
+            for i, class_utilities in enumerate(utilities):
+                class_probabilities = tf_ops.softmax_with_availabilities(
+                    items_logit_by_choice=class_utilities,
+                    available_items_by_choice=available_items_by_choice,
+                    normalize_exit=self.add_exit_choice,
+                    axis=-1,
+                )
+                probabilities.append(class_probabilities * latent_probabilities[i])
+            # Summing over the latent classes
+            probabilities = tf.reduce_sum(probabilities, axis=0)
             # Negative Log-Likelihood
             neg_loglikelihood = self.loss(
                 y_pred=probabilities,
                 y_true=tf.one_hot(choices, depth=probabilities.shape[1]),
[…]
