+++ choice_learn/models/latent_class_base_model.py
@@ -1,8 +1,9 @@
 """Base class for latent class choice models."""
 
+import time
+
 import numpy as np
 import tensorflow as tf
-import time
 import tqdm
 
 import choice_learn.tf_ops as tf_ops
@@ -68,9 +69,7 @@ class BaseLatentClassModel(object):
         self.lr = lr
         self.batch_size = batch_size
 
-        self.loss = tf_ops.CustomCategoricalCrossEntropy(
-            from_logits=False, label_smoothing=0.0
-        )
+        self.loss = tf_ops.CustomCategoricalCrossEntropy(from_logits=False, label_smoothing=0.0)
         self.exact_nll = tf_ops.CustomCategoricalCrossEntropy(
             from_logits=False,
             label_smoothing=0.0,
@@ -84,17 +83,18 @@ class BaseLatentClassModel(object):
 
     @property
     def trainable_weights(self):
-        """Returns trainable weights.
+        """Return trainable weights.
 
         Returns
         -------
         list
-           list of trainable weights."""
+           list of trainable weights.
+        """
         weights = [self.latent_logits]
         for model in self.models:
             weights += model.trainable_weights
         return weights
-    
+
     def instantiate(self, **kwargs):
         """Instantiate the model."""
         init_logit = tf.Variable(
@@ -265,7 +265,7 @@ class BaseLatentClassModel(object):
                 else:
                     print(f"Optimizer {self.optimizer} not implemnted, switching for default Adam")
                     self.optimizer = tf.keras.optimizers.Adam(self.lr)
-                
+
             return self._fit_with_gd(
                 choice_dataset=choice_dataset, sample_weight=sample_weight, verbose=verbose
             )
@@ -555,7 +555,6 @@ class BaseLatentClassModel(object):
         grads = tape.gradient(neg_loglikelihood, self.trainable_weights)
         self.optimizer.apply_gradients(zip(grads, self.trainable_weights))
         return neg_loglikelihood
-    
 
     def _fit_with_gd(
         self,
@@ -640,7 +639,7 @@ class BaseLatentClassModel(object):
                     )
 
                     train_logs["train_loss"].append(neg_loglikelihood)
-                    temps_logs = {k: tf.reduce_mean(v) for k, v in train_logs.items()}
+                    # temps_logs = {k: tf.reduce_mean(v) for k, v in train_logs.items()}
                     # self.callbacks.on_train_batch_end(batch_nb, logs=temps_logs)
 
                     # Optimization Steps
@@ -676,7 +675,7 @@ class BaseLatentClassModel(object):
                         choices_batch,
                     )
                     train_logs["train_loss"].append(neg_loglikelihood)
-                    temps_logs = {k: tf.reduce_mean(v) for k, v in train_logs.items()}
+                    # temps_logs = {k: tf.reduce_mean(v) for k, v in train_logs.items()}
                     # self.callbacks.on_train_batch_end(batch_nb, logs=temps_logs)
 
                     # Optimization Steps
@@ -728,7 +727,7 @@ class BaseLatentClassModel(object):
                         )[0]["optimized_loss"]
                     )
                     val_logs["val_loss"].append(test_losses[-1])
-                    temps_logs = {k: tf.reduce_mean(v) for k, v in val_logs.items()}
+                    # temps_logs = {k: tf.reduce_mean(v) for k, v in val_logs.items()}
                     # self.callbacks.on_test_batch_end(batch_nb, logs=temps_logs)
 
                 test_loss = tf.reduce_mean(test_losses)
@@ -740,7 +739,7 @@ class BaseLatentClassModel(object):
                 ]
                 train_logs = {**train_logs, **val_logs}
 
[…]
