### 2024-08-26 940564a
FIX: optimizer handler to be able to continue model.fit with GD
