### 2024-08-26 082deee
Fix EM algorithm to fit Latent Class model (#152)
