+++ choice_learn/models/latent_class_base_model.py
@@ -1,5 +1,7 @@
 """Base class for latent class choice models."""
 
+import time
+
 import numpy as np
 import tensorflow as tf
 import tqdm
@@ -17,6 +19,7 @@ class BaseLatentClassModel(object):
         model_parameters,
         fit_method,
         epochs,
+        batch_size=128,
         optimizer=None,
         add_exit_choice=False,
         tolerance=1e-6,
@@ -34,6 +37,7 @@ class BaseLatentClassModel(object):
             hyper-parameters of the models
         fit_method : str
             Method to estimate the parameters: "EM", "MLE".
+            "EM" for Expectation-Maximization, "MLE" for Maximum Likelihood Estimation
         epochs : int
             Number of epochs to train the model.
         optimizer: str, optional
@@ -64,14 +68,38 @@ class BaseLatentClassModel(object):
         self.tolerance = tolerance
         self.optimizer = optimizer
         self.lr = lr
+        self.batch_size = batch_size
 
-        self.loss = tf_ops.CustomCategoricalCrossEntropy(from_logits=False, label_smoothing=0)
+        self.loss = tf_ops.CustomCategoricalCrossEntropy(from_logits=False, label_smoothing=0.0)
+        self.exact_nll = tf_ops.CustomCategoricalCrossEntropy(
+            from_logits=False,
+            label_smoothing=0.0,
+            sparse=False,
+            axis=-1,
+            epsilon=1e-10,
+            name="exact_categorical_crossentropy",
+            reduction="sum_over_batch_size",
+        )
         self.instantiated = False
+
+    @property
+    def trainable_weights(self):
+        """Return trainable weights.
+
+        Returns
+        -------
+        list
+           list of trainable weights.
+        """
+        weights = [self.latent_logits]
+        for model in self.models:
+            weights += model.trainable_weights
+        return weights
 
     def instantiate(self, **kwargs):
         """Instantiate the model."""
         init_logit = tf.Variable(
-            tf.random_normal_initializer(0.0, 0.02, seed=42)(shape=(self.n_latent_classes - 1,)),
+            tf.random_normal_initializer(0.0, 0.08, seed=42)(shape=(self.n_latent_classes - 1,)),
             name="Latent-Logits",
         )
         self.latent_logits = init_logit
@@ -123,10 +151,7 @@ class BaseLatentClassModel(object):
             choices,
         )
 
-        latent_probabilities = tf.concat(
-            [[tf.constant(1.0)], tf.math.exp(self.latent_logits)], axis=0
-        )
-        latent_probabilities = latent_probabilities / tf.reduce_sum(latent_probabilities)
+        latent_probabilities = self.get_latent_classes_weights()
         # Compute probabilities from utilities & availabilties
         probabilities = []
         for i, class_utilities in enumerate(utilities):
@@ -148,7 +173,7 @@ class BaseLatentClassModel(object):
                 y_true=tf.one_hot(choices, depth=probabilities.shape[1]),
                 sample_weight=sample_weight,
             ),
-            "NegativeLogLikelihood": tf.keras.losses.CategoricalCrossentropy()(
+            "NegativeLogLikelihood": self.exact_nll(
                 y_pred=probabilities,
                 y_true=tf.one_hot(choices, depth=probabilities.shape[1]),
                 sample_weight=sample_weight,
@@ -227,12 +252,22 @@ class BaseLatentClassModel(object):
             )
 
         if self.fit_method.lower() == "mle":
-            if self.optimizer.lower() == "lbfgs" or self.optimizer.lower() == "l-bfgs":
-                return self._fit_with_lbfgs(
-                    choice_dataset=choice_dataset, sample_weight=sample_weight, verbose=verbose
-                )
[…]
