### 2024-08-26 0a67406
Fix logos on doc landing page (#153)
* FIX: logos sizes in mkdocs landing page
