### 2024-09-06 #6899 @crvernon
:wave: @VincentAuriau - this is making good progress.  Here are a few comments that I need you to address:

- [ ] In your acknowledgement in the paper, "...providing of the proprietary..." should be written "...provision of the proprietary..."

So here is what we have left to do:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list and order (should match who is on the paper exactly). You may also add the authors' ORCID.

- [ ]  Please respond with the DOI and the archived version number here
