### 2024-09-06 1d742ec
Change version to v1.0.0
