### 2024-09-10 #6899 @VincentAuriau
Hello @crvernon,

I made the changes you asked for as well as the release and archived it in Zenodo.
Here is the DOI: [10.5281/zenodo.13712222](https://doi.org/10.5281/zenodo.13712222), the metadata should match the paper's ones.
