+++ docs/paper/paper.md
@@ -55,18 +55,18 @@ The main contributions are summarized in Tables \ref{tab:comparison1} and \ref{t
 
 | Package      | Data Format      |  Data Batching     | Assortment   | Pricing        |
 | :----------: | :--------------: | :----------------: | :----------: | :------------: |
-| Biogeme      | wide             | $\times$           |   $\times$   |   $\times$     |
-| PyLogit      | long             |  $\times$          |   $\times$   |   $\times$     |
-| Torch-Choice | Multi Index      | $\checkmark$       |   $\times$   |   $\times$     |
+| Biogeme      | wide             | $\times$           | $\times$     | $\times$       |
+| PyLogit      | long             | $\times$           | $\times$     | $\times$       |
+| Torch-Choice | Multi Index      | $\checkmark$       | $\times$     | $\times$       |
 | Choice-Learn | Features Storage | $\checkmark$       | $\checkmark$ | $\checkmark$   |
 
 
 : Comparison of the different packages for modelization. CondL, NestL, MixL, and LatC respectively indicate the Conditional Logit, Nested Logit, Mixed Logit and Latent Class models.\label{tab:comparison2}
 
-|   Package    | Traditional Models              | NeuralNet Models   | Custom Models  | Non-Stochastic Optimizer |  Stochastic Optimizer  |
+|   Package    | Traditional Models              | NeuralNet Models   | Custom Models  | Non-Stochastic Optimizer | Stochastic Optimizer |
 | :----------: | :-----------------------------: | :----------------: | :------------: | :----------------------: | :------------------: |
 | Biogeme      | CondL, NestL,MixL, LatC & more  | $\times$           | $\checkmark$   | Newton BFGS              | $\quad \times \quad$ |
-| PyLogit      | CondL, NestL,MixL,  Asymmetric  |  $\times$          | $\times$       | BFGS                     |   $\times$           |
+| PyLogit      | CondL, NestL,MixL, Asymmetric   | $\times$           | $\times$       | BFGS                     | $\times$             |
 | Torch-Choice | CondL, NestL                    | $\times$           | $\times$       | L-BFGS                   | $\checkmark$         |
 | Choice-Learn | CondL, NestL, LatC              | $\checkmark$       | $\checkmark$   | L-BFGS                   | $\checkmark$         |
 
