### 2024-09-10 #6899 @VincentAuriau
@crvernon from my understanding the tables in the paper.md file were creating this issue. I hopefully fixed it with recent commits.
