### 2024-09-10 #6899 @VincentAuriau
@tmigot @samuelduchesne @crvernon Thank you so much for taking the time to review the paper and the package as well as for your valuable feedbacks !
